+++ ndk-build/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- Added support for adding extra files to the APK.
+
 # 0.1.4 (2020-11-25)
 
 - On Windows, fixed UNC path handling for resource folder